improvement(radiocomponent): update directory name and format story
